merge

+++ src/Hangfire.SqlServer/IPersistentJobQueueMonitoringApi.cs
@@ -21,8 +21,12 @@ namespace Hangfire.SqlServer
     public interface IPersistentJobQueueMonitoringApi
     {
         IEnumerable<string> GetQueues();
-        IEnumerable<int> GetEnqueuedJobIds(string queue, int from, int perPage);
-        IEnumerable<int> GetFetchedJobIds(string queue, int from, int perPage);
+
+        IEnumerable<long> GetEnqueuedJobIds(string queue, int from, int perPage);
+
+        // TODO: Extend return type by including DateTime to allow getting the FetchedAt value in 2.0
+        IEnumerable<long> GetFetchedJobIds(string queue, int from, int perPage);
+
         EnqueuedAndFetchedCountDto GetEnqueuedAndFetchedCount(string queue);
     }
 }
